requirements.txt
alembic==1.16.4
app-scheduler==0.0.1
APScheduler==3.11.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.2
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.2
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jira==3.10.5
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
py3db==0.0.19
pycparser==2.23
pymongo==4.15.3
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==22.0.0
